Move private structs into dedicated header
